+++ locales/en.json
@@ -1305,6 +1305,8 @@
     "fixedPrice": "Fixed Price",
     "collectionFloorPrice": "Collection Floor Price",
     "potentialEarnings": "Potential Earnings",
-    "listForSale": "List For Sale"
+    "listForSale": "List For Sale",
+    "chooseAPrice": "Choose a price",
+    "floorPrice": "Floor Price"
   }
 }
